Working push.

+++ src/preprocess.py
@@ -11,7 +11,7 @@ from tqdm import tqdm
 from typing import Optional
 
 
-def preprocess(fname: str, data_dir: Optional[str] = "data") -> None:
+def preprocess_fn(fname: str, data_dir: Optional[str] = "data") -> None:
     '''
     Preprocess fname, which has the structure:
     {"indoml_id": <id>, "title": <title>, "store": <store>, "details_Manufacturer": ...}
@@ -42,6 +42,9 @@ def preprocess(fname: str, data_dir: Optional[str] = "data") -> None:
     else:
         out_name = fname.split('.')[0].split('_')[1]
     output_file = f"{data_dir}/{out_name}.json"
+    if os.path.exists(output_file):
+        print(f"{output_file} already exists, skipped.")
+        return
     print(f"Preprocessing {file}, will store in {output_file} ...")
     with open(file, 'r') as f:
         for line in tqdm(f):
@@ -83,7 +86,7 @@ def categorize(attr_name: str = "details_Brand") -> None:
     
     # get unique values
     col_vals = set()
-    data = json.load(open(f"data/test_sol.json", 'r'))
+    data = json.load(open(f"data/val_sol.json", 'r'))
     for item in data:
         col_vals.add(item[attr_name])
     
@@ -92,11 +95,15 @@ def categorize(attr_name: str = "details_Brand") -> None:
     # revert the dictionary
     idx2col = {idx: col for col, idx in col2idx.items()}
     # save
+    if os.path.exists(f'data/maps/{attr_name}2idx.pkl'):
+        print(f"Maps for {attr_name} already exist, skipped.")
+        return
     with open(f'data/maps/{attr_name}2idx.pkl', 'wb') as f:
         pickle.dump(col2idx, f)
     with open(f'data/maps/idx2{attr_name}.pkl', 'wb') as f:
         pickle.dump(idx2col, f)
 
+    print(f'Categorized {col}.')
     return
 
 
@@ -109,13 +116,11 @@ if __name__ == '__main__':
     
     args = parser.parse_args()
     
-    preprocess("attrebute_test.data", args.data_dir)
-    preprocess("attrebute_test.solution", args.data_dir)
-    preprocess("attrebute_val.data", args.data_dir)
-    preprocess("attrebute_val.solution", args.data_dir)
-    preprocess("attrebute_train.data", args.data_dir)
-    preprocess("attrebute_train.solution", args.data_dir)
+    preprocess_fn("attrebute_test.data", args.data_dir)
+    preprocess_fn("attrebute_val.data", args.data_dir)
+    preprocess_fn("attrebute_val.solution", args.data_dir)
+    preprocess_fn("attrebute_train.data", args.data_dir)
+    preprocess_fn("attrebute_train.solution", args.data_dir)
     
     for col in ['details_Brand', 'L0_category', 'L1_category', 'L2_category', 'L3_category', 'L4_category']:
         categorize(col)
-        print(f'{col} done.')
